requirements.txt
Django==5.0.6
PyMySQL==1.1.1
pdf2image==1.17.0
Pillow>=11.0.0
python-dotenv==1.0.1
whitenoise==6.6.0